Rename http.util -> http._compat
